requirements.txt
Flask==2.3.3
black==23.9.1
openai==0.28.0
python-dotenv==1.0.0
